+++ example/vent.py
@@ -1,13 +1,13 @@
 import cadquery as cq
 from cqdome import greeble
 
-vent = greeble.vent_hexagon(
-    radius=58,
-    height=4,
-    frame_size = 9,
-    vent_width = 2,
-    vent_space = 0.6,
-    vent_rotate=45
-)
-
-cq.exporters.export(vent,'stl/vent_hexagon.stl')
+bp = greeble.VentHexagon()
+bp.radius = 58
+bp.height = 4
+bp.frame_size = 9
+bp.vent_width = 2
+bp.vent_space = 0.6
+bp.vent_rotate = 45
+bp.make()
+vent = bp.build()
+cq.exporters.export(vent,'stl/VentHexagon.stl')