+++ src/cqdome/greeble/CutKeyHexagon.py
@@ -1,0 +1,7 @@
+import cadquery as cq
+from . import HexagonBase, make_hexagon
+
+class CutKeyHexagon(HexagonBase)
+    def __init__(self):
+        print('CutKeyHexagon __init__')
+        super().__init__(self)