+++ src/cqdome/greeble/CutKeyPentagon.py
@@ -1,0 +1,7 @@
+import cadquery as cq
+from . import PentagonBase, make_pentagon
+
+class CutKeyPentagon(PentagonBase)
+    def __init__(self):
+        print('CutKeyPentagon __init__')
+        super().__init__(self)