+++ src/cqdome/greeble/__init__.py
@@ -1,2 +1,8 @@
-from .vent_hexagon import vent_hexagon
+from .BaseHexagon import BaseHexagon, make_hexagon
+from .BasePentagon import BasePentagon, make_pentagon
 from .DoorHexagon import DoorHexagon
+from .VentHexagon import VentHexagon
+
+
+#from .CutKeyHexagon import CutKeyHexagon
+#from .CutKeyPentagon import CutKeyPentagon