+++ changes.md
@@ -1,7 +1,18 @@
 # CQDome changelog
 
 ## Main wip
-* Added Vent Greeble
+* Updated Dependencies
+* Added BaseHexagon
+* Added BasePentagon
+* Refactored vent_hexagon to VentHexagon
+* Cleaned up imports
+* Added door.py example
+
+## 0.0.2
+* Refactored Dome code to version 2
+* Added Vent greeble
+* Added Door greeble
+
 
 ## 0.0.1
 * Initial release